perf: Implement comprehensive training efficiency optimizations
This commit implements a complete suite of training optimizations including
vectorization, mixed precision, and memory optimizations. All changes are
verified with comprehensive tests and benchmarks.

## 1. Vectorized AR(1) Cholesky Decomposition

**Performance:** 435x speedup (56.5ms → 0.13ms for H=90)

- Replaced O(H²) Python loops with O(H) vectorized torch operations
- Uses closed-form AR(1) Cholesky formula with torch.arange, torch.pow, torch.tril_indices
- Verified correctness: max diff < 1µs vs torch.linalg.cholesky
- Kept original loop-based implementation as _build_ar1_cholesky_direct_loop() for reference
- Files: vae/full_covariance_prior.py

## 2. Vectorized KL Divergence Computation

**Performance:** Computes all latent dimensions in parallel (no more Python loops)

- Replaced dimension loop with einsum operations for parallel computation
- Vectorized: trace term, Mahalanobis distance, log det ratio
- Verified correctness: diff < 1e-4 relative error vs loop version
- Kept original loop-based implementation as _kl_divergence_full_covariance_loop() for reference
- Files: vae/full_covariance_prior.py

## 3. FP32 Default Precision

**Expected benefits:** 2x memory reduction, 2-3x speedup on modern GPUs

- Changed all dataset defaults: torch.float64 → torch.float32
- Updated 17 training/generation scripts to use FP32 by default
- Verified training stability: 5-epoch test shows stable convergence (no NaN)
- Kept intentional FP64 usage (GMM fitting) unchanged for numerical stability
- Files: vae/datasets.py, vae/datasets_randomized.py, param_search.py,
  train_quantile_models.py, generate_*.py, main_analysis.py, experiments/backfill/**

## 4. Memory Optimizations

- Added zero_grad(set_to_none=True) in CVAEFullCovPrior for faster gradient reset
- Files: vae/cvae_full_cov_prior.py

## 5. Test Infrastructure

**New files:**
- test_code/test_optimizations.py (12 correctness tests)
  - Cholesky vectorization tests (5 tests)
  - KL divergence vectorization tests (5 tests)
  - FP32 conversion tests (2 tests)

- test_code/test_optimization_benchmarks.py (performance benchmarks)
  - Cholesky construction benchmark
  - KL divergence benchmark
  - GPU memory usage benchmark

## Verification

✅ All 47 V4 tests pass (test_cvae_full_cov_prior.py, test_full_covariance_prior.py)
✅ All 12 optimization tests pass (test_optimizations.py)
✅ FP32 training stability verified (5-epoch test: loss 0.181 → 0.017)
✅ Vectorized implementations verified correct (max diff < 1µs for Cholesky, < 1e-4 for KL)

+++ experiments/backfill/context60/train_backfill_context60.py
@@ -444,7 +444,7 @@ def train_phase4_ar_h90(model, optimizer, train_loader, valid_loader,
 def main():
     # Setup
     set_seeds(0)
-    torch.set_default_dtype(torch.float64)
+    torch.set_default_dtype(torch.float32)
 
     cfg = BackfillContext60Config
     output_dir = cfg.checkpoint_dir
